requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
dataclasses-json==0.6.6
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-jwt==1.19.2
drf-spectacular==0.27.2
frozenlist==1.4.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.25
langchain-openai==0.1.19
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.0
langsmith==0.1.94
lz4==4.3.3
Markdown==3.6
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.3
packaging==23.2
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.2
rpds-py==0.18.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
